Add ActivityType ListView
We add activity_type ListView that display the activity_types with their total activities' duration.
This total activities' duration might need to be explained:
Each Activity belongs to a specific activity_type and has two time field 'time_from and time_to'. Now, with this commit, we add a duration field(a Float type field) to the Activity model that store the duration between the two times in hours. And each activity_type sum the durations of all its activities; that(the sum) what will be displayed on the ListView.