Update activities/diary_list.html Template
Update the activities/diary_list.html to display AFTER_MIDNIGHT sleep duration. To do that we added a method, get_after_midnight_sleep_duration, to the Diary model. The method does the calculation; because the spleeping time of this period is not registered directly.

We also added Docstrings to the Diary model, ActivityAutocomplete, and ActivityTypeAutocomplete that describe some of their works.

With this commit we also starting writing tests. We added a single test, test_activity_duration_should_be_positive.