Update ActivityType validators and add a corresponding test
We are continuing in making the ActivityType validation more robust. We have a validator, validate_activity_type_form, that call other subvalidators. with this commit we added _validate_activity_type_reciting_quran_form subvalidator.

We also changed the many to many field, market, to be many to one field(foreign key field).