Squash activities app migrations
We want to add another application to our project so we see that it is better to squash the current app(activities) migrations into a single file.
Because we have data to insert into the database during the migration we end up with two migration files; one for the whole schema squashed in a single file and the other is for inserting the data.